Add intrusive_list and intrusive_stack data-structures.
The intrusive_list differs from intrusive_queue in that it
supports removal of elements in the middle, which is
necessary for implementing cancellation.

The intrusive_stack allows only pushing/popping from
the front but only requires storing a single intrusive
pointer. Great for free-lists.

Minor refactoring of atomic_intrusive_queue to add
the ability to mark the queue as active without
dequeueing the items.

Also added ability to dequeue items in reverse as
a stack, which avoids needing to traverse the list
twice if you were already going to traverse the list.